Add files via upload
The goal of this project is to analyze Skopje's air pollution patterns in order to better understand the environmental issues facing the city and develop mitigation plans. The project will identify high-risk locations, analyze the impact of pollution on public health, and assess the efficacy of current regulatory measures using sophisticated data analytics and visualization approaches. The research results will be utilized to produce practical suggestions that legislators, urban planners, and community members may implement to tackle air quality problems and encourage sustainable growth in Skopje.